refactor: isolate openai streaming
