+++ scripts/rust-source-size-baseline.txt
@@ -1,6 +1,6 @@
 # Rust source files above the 1,000-line target. Keep sorted by path.
 # A file may only stay equal to this maximum or shrink with this entry tightened.
-crates/llm/src/openai_chat_completions.rs 2244
+crates/llm/src/openai_chat_completions.rs 1447
 crates/llmfs/src/lib.rs 1797
 crates/service-manager/src/package.rs 2947
 crates/service-manager/src/runtime.rs 2243
